+++ src/main/resources/archetype-resources/.gitlab-ci.yml
@@ -36,7 +36,7 @@ services:
 CucumberBackendIT:
   stage: test
   script:
-    - 'mvn verify -Dit.test=${groupId}.${artifactId}.runner.CucumberBackendIT'
+    - 'mvn verify -Dit.test=${groupId}.${artifactId}.runners.CucumberBackendIT'
   artifacts:
     paths:
     - target/*.json
@@ -46,7 +46,7 @@ CucumberBackendIT:
 CucumberRestIT:
   stage: test
   script:
-    - 'mvn verify -Dit.test=${groupId}.${artifactId}.runner.CucumberRestIT'
+    - 'mvn verify -Dit.test=${groupId}.${artifactId}.runners.CucumberRestIT'
   artifacts:
     paths:
     - target/*.json
@@ -56,7 +56,7 @@ CucumberRestIT:
 CucumberSeleniumIT:
   stage: test
   script:
-    - 'mvn verify -Dit.test=${groupId}.${artifactId}.runner.CucumberSeleniumIT -DSELENIUM_NODE=node:4444'
+    - 'mvn verify -Dit.test=${groupId}.${artifactId}.runners.CucumberSeleniumIT -DSELENIUM_NODE=node:4444'
   artifacts:
     paths:
     - target/executions/**
