+++ src/main/resources/archetype-resources/src/test/java/__groupId__/__artifactId__/runners/CucumberBackendIT.java
@@ -1,7 +1,7 @@
-package ${groupId}.${artifactId}.runner;
+package ${groupId}.${artifactId}.runners;
 
 import cucumber.api.CucumberOptions;
-import ${groupId}.${artifactId}.hooks.BaseTest;
+import ${groupId}.${artifactId}.utils.BaseTest;
 
 
 /**
@@ -12,14 +12,15 @@ import ${groupId}.${artifactId}.hooks.BaseTest;
  * features
  */
 @CucumberOptions(plugin = {
-                "json:target/CucumberRestIT.json",
-                "html:target/CucumberRestIT",
-                "junit:target/CucumberRestIT.xml"
+                "json:target/CucumberBackendIT.json",
+                "html:target/CucumberBackendIT",
+                "junit:target/CucumberBackendIT.xml"
         }, features =
         {
-                "src/test/resources/features/cucumber_rest_test.feature"
+                "src/test/resources/features/cucumber_backend_test.feature"
         },
         glue = "classpath:${groupId}.${artifactId}.glue")
-public class CucumberRestIT extends BaseTest {
+public class CucumberBackendIT extends BaseTest {
+
 
 }
