+++ src/main/resources/archetype-resources/src/test/resources/features/cucumber_backend_test.feature
@@ -4,7 +4,7 @@ Feature: Testing backend related functionality
   related to the use of variables, JSON replacing, etc).
 
   Execute this test like this:
-  mvn verify -Dit.test=${groupId}.${artifactId}.runner.CucumberBackendIT
+  mvn verify -Dit.test=${groupId}.${artifactId}.runners.CucumberBackendIT
 
   You can check a more complete list of steps and examples at:
   https://github.com/vpTechOSS/gingerspec/wiki/Gherkin-extended-language
